loadXDF
src directory includes Yida's code on loading XDF files. Most of Yida's
files are in file_handling_impl directory, with some slight changes in
src.pro as well. Yida also added "XDF" to the enum in biosig.h in order
to let sigviewer recognize XDF

+++ src/gui_impl/commands/zoom_gui_command.h
@@ -1,22 +1,3 @@
-// Copyright 2013 Clemens Brunner, Thomas Brunner, Christoph Eibel,
-// Alois Schlögl, Oliver Terbu.
-
-// This file is part of SigViewer.
-//
-// SigViewer is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// SigViewer is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with SigViewer. If not, see <http://www.gnu.org/licenses/>.
-
-
 #ifndef ZOOM_GUI_COMMAND_H
 #define ZOOM_GUI_COMMAND_H
 
@@ -76,8 +57,8 @@ private:
     //-------------------------------------------------------------------------
     float32 minPixelPerSample ();
 
-    static int const ZOOM_FACTOR_ = 2;
-    static int const MAX_HORIZONTAL_ZOOM_IN_ = 32;
+    static float const ZOOM_FACTOR_ = 2;
+    static float const MAX_HORIZONTAL_ZOOM_IN_ = 32;
 
     static QString const GOTO_;
     static QString const ZOOM_IN_VERTICAL_;
